+++ trainingAnalysisTool_library/lib/running.hpp
@@ -13,7 +13,7 @@ class RUNNING : public TRAININGINSTANCE
         std::map<std::string, double> m_heartRateZones; //Might not be private to RUNNING if I add more subclasses which uses heart rate zones
 
     public:
-        RUNNING(std::string type, double duration, double week, double cadence, std::map<std::string, double> heartRateZones)
+        RUNNING(std::string type, double duration, int week, double cadence, std::map<std::string, double> heartRateZones)
             : TRAININGINSTANCE{type, duration, week}, m_cadance{cadence}, m_heartRateZones{heartRateZones}
             {
                 TRAININGINSTANCE::setCadenceUsedTrue();
